pyproject.toml
[tool.poetry]
name = "kubera-server"
version = "0.1.0"
description = ""
authors = ["Jath Palasubramaniam <jathpala@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"
pydantic-settings = "^2.3.4"
taskipy = "^1.13.0"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
alembic = "^1.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
dev = "fastapi dev kubera_server"
